#2999 Modify formatting and function per PR feedback

+++ service/src/main/java/skills/services/admin/ProjAdminService.groovy
@@ -245,6 +245,11 @@ class ProjAdminService {
             throw new SkillException("Project with id [${projectId}] cannot be deleted as it is currently referenced by one or more global badges")
         }
 
+        SettingsResult isDeleteProtected = settingsService.getProjectSetting(projectId, "project-deletion-protection")
+        if (isDeleteProtected.value == "true") {
+            throw new SkillException("Project [${projectId}] cannot be deleted as it has deletion protection enabled")
+        }
+
         List<SkillDef> childSkills = skillDefRepo.findAllByProjectIdAndType(projectId, SkillDef.ContainerType.Skill)
         childSkills.each {
             skillsAdminService.removeCatalogImportedSkills(it)
@@ -377,7 +382,7 @@ class ProjAdminService {
     private  List<ProjectResult> loadProjectsForRoot(Map<String, Integer> projectIdSortOrder, String userId, Boolean isNotCommunityMember) {
         List<SettingsResult> pinnedProjectSettings = settingsService.getUserProjectSettingsForGroup(userId, rootUserPinnedProjectGroup)
         List<String> pinnedProjects = pinnedProjectSettings.collect { it.projectId }
-        List<SettingsResult> projectSettings = settingsService.getProjectSettingForAllProjectsInList("project-protection", pinnedProjects)
+        List<SettingsResult> projectSettings = settingsService.getProjectSettingForAllProjectsInList("project-deletion-protection", pinnedProjects)
 
         List<ProjSummaryResult> projects = projDefRepo.getAllSummariesByProjectIdIn(pinnedProjects)
         if (isNotCommunityMember) {
@@ -460,7 +465,7 @@ class ProjAdminService {
                 projects = projects.findAll { !it.protectedCommunityEnabled}
             }
             List<String> projectIds = projects.collect{ it.projectId }
-            List<SettingsResult> projectSettings = settingsService.getProjectSettingForAllProjectsInList("project-protection", projectIds)
+            List<SettingsResult> projectSettings = settingsService.getProjectSettingForAllProjectsInList("project-deletion-protection", projectIds)
             finalRes = projects?.unique({ it.projectId })?.collect({
                 ProjectResult res = convert(it, projectIdSortOrder)
                 SettingsResult isProtected = projectSettings.find{ setting -> setting.projectId == it.projectId }
